moves compat fixes
