+++ Fakebook.Posts/Fakebook.Posts.DataAccess/PostsRepository.cs
@@ -1,8 +1,14 @@
 using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using Fakebook.Posts.Domain;
+using Fakebook.Posts.Domain.Models;
 
 namespace Fakebook.Posts.DataModel
 {
-    public class PostsRepository 
+    public class PostsRepository : IPostsRepository
     {
         private readonly FakebookPostsContext _context;
         public PostsRepository(FakebookPostsContext context)
@@ -14,12 +20,12 @@ namespace Fakebook.Posts.DataModel
 
         public bool IsReadOnly => false;
 
-        public void Add(Fakebook.Posts.Domain.Post item)
+        public void Add(Fakebook.Posts.Domain.Models.Post item)
         {
             throw new System.NotImplementedException();
         }
 
-        public ValueTask<bool> AddAsync(Fakebook.Posts.Domain.Post post)
+        public ValueTask<Fakebook.Posts.Domain.Models.Post> AddAsync(Fakebook.Posts.Domain.Models.Post post)
         {
             throw new System.NotImplementedException();
         }
@@ -29,34 +35,39 @@ namespace Fakebook.Posts.DataModel
             throw new System.NotImplementedException();
         }
 
-        public bool Contains(Fakebook.Posts.Domain.Post item)
+        public bool Contains(Fakebook.Posts.Domain.Models.Post item)
         {
             throw new System.NotImplementedException();
         }
 
-        public void CopyTo(Fakebook.Posts.Domain.Post[] array, int arrayIndex)
+        public void CopyTo(Fakebook.Posts.Domain.Models.Post[] array, int arrayIndex)
         {
             throw new System.NotImplementedException();
         }
 
-        public IAsyncEnumerator<Fakebook.Posts.Domain.Post> GetAsyncEnumerator(CancellationToken cancellationToken = default)
+        IAsyncEnumerator<Post> GetAsyncEnumerator(CancellationToken cancellationToken = default)
         {
             throw new System.NotImplementedException();
         }
 
-        public IEnumerator<Fakebook.Posts.Domain.Post> GetEnumerator()
+        public bool Remove(Fakebook.Posts.Domain.Models.Post item)
         {
             throw new System.NotImplementedException();
         }
 
-        public bool Remove(Fakebook.Posts.Domain.Post item)
+        IAsyncEnumerator<Post> IAsyncEnumerable<Post>.GetAsyncEnumerator(CancellationToken cancellationToken)
         {
-            throw new System.NotImplementedException();
+            throw new NotImplementedException();
         }
 
         IEnumerator IEnumerable.GetEnumerator()
         {
             throw new System.NotImplementedException();
+        }
+
+        IEnumerator<Post> IEnumerable<Post>.GetEnumerator()
+        {
+            throw new NotImplementedException();
         }
     }
 }
