+++ Fakebook.Posts/Fakebook.Posts.Domain/IPostsRepository.cs
@@ -6,7 +6,7 @@ using Fakebook.Posts.Domain.Models;
 namespace Fakebook.Posts.Domain {
     public interface IPostsRepository : ICollection<Post>, IAsyncEnumerable<Post>
     {
-        public ValueTask<Post> AddAsync(Post post);
-        // public ValueTask<bool> RemoveAsync(Post post);
+        ValueTask<Post> AddAsync(Post post);
+        // ValueTask<bool> RemoveAsync(Post post);
     }
 }
