+++ Fakebook.Posts/Fakebook.Posts.Domain/Models/Comment.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace Fakebook.Posts.Domain {
+namespace Fakebook.Posts.Domain.Models {
     public class Comment {
     }
 }
